requirements.txt
openai
python-dotenv
pyyaml
jsonschema
streamlit
